requirements_venv_rlbench.txt
# RAEDiTRobotics — venv_rlbench requirements
# Install: pip install -r requirements_venv_rlbench.txt
#
# Prerequisites (install separately):
#   - PyTorch 2.9+ with CUDA 12 (https://pytorch.org)
#   - PyRep: pip install git+https://github.com/stepjam/PyRep.git
#   - RLBench: pip install -e git+https://github.com/stepjam/RLBench.git#egg=rlbench
#   - robomimic: pip install -e git+https://github.com/ARISE-Initiative/robomimic.git#egg=robomimic
#   - robosuite 1.5.2: pip install robosuite==1.5.2

# Core ML
diffusers>=0.37.0
einops>=0.8.0
timm>=1.0.0
transformers>=5.0.0
safetensors>=0.7.0

# Vision / image processing
opencv-python-headless>=4.8.0
pillow>=10.0.0
imageio>=2.30.0
imageio-ffmpeg>=0.6.0
lpips>=0.1.4

# Data / numerics
h5py>=3.10.0
numpy>=1.26.0,<2.0
scipy>=1.14.0
pyquaternion>=0.9.9

# Training utilities
tqdm>=4.65.0
tensorboard>=2.15.0
tensorboardX>=2.6.0

# Config / logging
PyYAML>=6.0
rich>=13.0.0

# Testing
pytest>=8.0.0

# Misc
matplotlib>=3.8.0
